projects/ai-trip-planner/requirements.txt
langchain
langgraph
langchain-community
langchain-experimental
langchain-google-community
langchain-google-community[places]
langchain_tavily
langchain_groq
langchain_openai
fastapi
python-dotenv
streamlit
uvicorn
httpx
requests
azure-identity

-e .